chore: Refactored and renamed test/controller/attachment_controller_test.rb to test/integration/attachments_preview_test.rb
